requirements.txt
fastapi==0.100.1
uvicorn==2.0.1
langchain==0.0.251
sentence-transformers==2.2.2